add sub category fragment
