--- app/src/main/res/drawable/cuisine_gradiant.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<shape xmlns:android="http://schemas.android.com/apk/res/android"
-    android:shape="oval">
-    <gradient
-        android:angle="90"
-        android:centerColor="#29000000"
-        android:endColor="#29000000"
-        android:startColor="#000000" />
-</shape>
