--- app/src/main/res/drawable/fav.xml
@@ -1,12 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="14dp"
-    android:height="13dp"
-    android:viewportWidth="14"
-    android:viewportHeight="13">
-  <path
-      android:strokeWidth="1"
-      android:pathData="M6.798,1.717L7.178,2.165L7.559,1.717C8.194,0.972 9.17,0.5 10.178,0.5C11.956,0.5 13.345,1.889 13.345,3.667C13.345,4.762 12.857,5.796 11.896,6.971C10.929,8.153 9.536,9.417 7.809,10.983L7.808,10.984L7.177,11.558L6.548,10.99C6.548,10.99 6.548,10.99 6.548,10.99C4.821,9.421 3.428,8.155 2.461,6.972C1.5,5.796 1.012,4.762 1.012,3.667C1.012,1.889 2.401,0.5 4.178,0.5C5.187,0.5 6.163,0.972 6.798,1.717ZM2.909,6.987C3.852,8.089 5.193,9.304 6.767,10.729L6.825,10.787L7.178,11.14L7.532,10.787L7.59,10.729C9.164,9.304 10.505,8.089 11.448,6.987C12.389,5.888 13.012,4.815 13.012,3.667C13.012,2.057 11.788,0.833 10.178,0.833C9.08,0.833 8.001,1.462 7.486,2.407H6.876C6.357,1.463 5.278,0.833 4.178,0.833C2.569,0.833 1.345,2.057 1.345,3.667C1.345,4.815 1.968,5.888 2.909,6.987Z"
-      android:fillColor="#ffffff"
-      android:fillAlpha="0.87"
-      android:strokeColor="#512C08"/>
-</vector>
