--- app/src/main/res/drawable/time.xml
@@ -1,6 +1,0 @@
-<vector android:height="24dp" android:viewportHeight="16"
-    android:viewportWidth="24" android:width="36dp" xmlns:android="http://schemas.android.com/apk/res/android">
-    <path android:fillColor="#512C08" android:pathData="M5.448,7.067L5.764,7.428C6.45,6.826 7.225,6.34 8.067,5.984L7.879,5.542C6.99,5.918 6.172,6.431 5.448,7.067L5.448,7.067Z"/>
-    <path android:fillColor="#512C08" android:pathData="M19.666,6.887C18.891,6.887 18.164,7.094 17.535,7.452C16.063,5.165 13.703,3.694 11.043,3.378C11.398,3.071 11.628,2.622 11.628,2.117C11.628,1.195 10.877,0.445 9.954,0.445C9.032,0.445 8.281,1.195 8.281,2.117C8.281,2.625 8.514,3.076 8.872,3.383C4.486,3.911 1.066,7.592 0.947,12.088L0,12.088V14.09H1.548L1.932,15.555H17.976L18.057,15.243C18.555,15.443 19.097,15.555 19.665,15.555C22.055,15.555 24,13.611 24,11.222C24,8.832 22.056,6.887 19.666,6.887ZM8.761,2.116C8.761,1.459 9.296,0.925 9.954,0.925C10.612,0.925 11.148,1.459 11.148,2.116C11.148,2.776 10.612,3.311 9.954,3.311C9.296,3.311 8.761,2.775 8.761,2.116ZM9.954,3.791C12.886,3.791 15.551,5.254 17.132,7.712C16.043,8.5 15.331,9.777 15.331,11.221C15.331,11.517 15.361,11.807 15.418,12.088L1.427,12.088C1.556,7.492 5.326,3.791 9.954,3.791ZM0.48,13.61V12.568H15.549C15.671,12.939 15.841,13.29 16.053,13.61H0.48ZM17.605,15.075H2.303L2.044,14.09L16.426,14.09C16.763,14.47 17.164,14.794 17.615,15.038L17.605,15.075ZM19.666,15.075C17.54,15.075 15.812,13.346 15.812,11.221C15.812,9.096 17.541,7.367 19.666,7.367C21.791,7.367 23.52,9.096 23.52,11.221C23.52,13.346 21.791,15.075 19.666,15.075Z"/>
-    <path android:fillColor="#512C08" android:pathData="M21.167,8.93L19.617,10.895L19.088,10.527L18.813,10.921L19.319,11.273L18.994,11.685L19.371,11.982L19.713,11.547L20.988,12.435L21.263,12.041L20.011,11.17L21.544,9.227L21.167,8.93Z"/>
-</vector>
