--- app/src/main/res/layout/fragment_cuisines.xml
@@ -1,17 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.core.widget.NestedScrollView xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    xmlns:tools="http://schemas.android.com/tools">
-        <GridLayout
-            android:id="@+id/gridLayout"
-            android:layout_width="wrap_content"
-            android:layout_height="match_parent"
-            android:layout_gravity="center_horizontal"
-            android:columnCount="2"
-            android:paddingBottom="80dp"
-            tools:context=".ui.CuisinesFragment">
-        </GridLayout>
-
-
-</androidx.core.widget.NestedScrollView>
