--- app/src/main/res/layout/item_cuisine.xml
@@ -1,31 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.cardview.widget.CardView xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="146dp"
-    android:layout_height="147dp"
-    android:layout_marginTop="8dp"
-    android:layout_marginEnd="42dp"
-    android:foreground="?android:attr/selectableItemBackground"
-    app:cardCornerRadius="100dp">
-
-    <ImageView
-        android:id="@+id/cuisineImageView"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:contentDescription="@string/category_image"
-        android:foreground="@drawable/cuisine_gradiant"
-        android:scaleType="fitXY"
-        android:src="@drawable/cuisine" />
-
-    <TextView
-        android:id="@+id/cuisineNameTextView"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_gravity="center_horizontal|bottom"
-        android:layout_marginBottom="8dp"
-        android:fontFamily="@font/poppins_medium"
-        android:textColor="@color/white"
-        android:textSize="14sp"
-        tools:text="Chinese" />
-</androidx.cardview.widget.CardView>
