--- app/src/main/res/layout/item_subcategory.xml
@@ -1,89 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.cardview.widget.CardView xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="162dp"
-    android:layout_height="200dp"
-    android:backgroundTint="#FFF8F7"
-    app:cardCornerRadius="12dp"
-    app:cardElevation="0dp">
-
-    <androidx.constraintlayout.widget.ConstraintLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent">
-
-        <androidx.constraintlayout.utils.widget.ImageFilterView
-            android:id="@+id/categoryImageView"
-            android:layout_width="match_parent"
-            android:layout_height="124dp"
-            android:scaleType="fitXY"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent"
-            app:round="12dp"
-            tools:src="@tools:sample/avatars" />
-
-        <androidx.cardview.widget.CardView
-            android:id="@+id/favoriteCardView"
-            android:layout_width="32dp"
-            android:layout_height="32dp"
-            android:layout_gravity="end"
-            android:layout_margin="8dp"
-            app:cardCornerRadius="100dp"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintTop_toTopOf="parent">
-
-            <ImageView
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:layout_gravity="center"
-                android:src="@drawable/fav" />
-        </androidx.cardview.widget.CardView>
-
-
-        <TextView
-            android:id="@+id/subCategoryTextView"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="8dp"
-            android:layout_marginTop="14dp"
-            android:textColor="#FF512C08"
-            android:textSize="16sp"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@id/categoryImageView"
-            tools:text="Kofta" />
-
-        <ImageView
-            android:id="@+id/cookTimeImageView"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginTop="4dp"
-            android:src="@drawable/time"
-            app:layout_constraintStart_toStartOf="@id/subCategoryTextView"
-            app:layout_constraintTop_toBottomOf="@id/subCategoryTextView" />
-
-        <TextView
-            android:id="@+id/cookTimeTextView"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="8dp"
-            android:textColor="#FF512C08"
-            android:textSize="14sp"
-            app:layout_constraintBottom_toBottomOf="@id/cookTimeImageView"
-            app:layout_constraintStart_toEndOf="@id/cookTimeImageView"
-            app:layout_constraintTop_toTopOf="@id/cookTimeImageView"
-            tools:text="15" />
-
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="4dp"
-            android:text="@string/min"
-            android:textColor="#FF512C08"
-            android:textSize="14sp"
-            app:layout_constraintBottom_toBottomOf="@id/cookTimeImageView"
-            app:layout_constraintStart_toEndOf="@id/cookTimeTextView"
-            app:layout_constraintTop_toTopOf="@id/cookTimeImageView" />
-
-    </androidx.constraintlayout.widget.ConstraintLayout>
-</androidx.cardview.widget.CardView>
